Revert "add version property to cycle and remove from scenario"
This reverts commit b80dcda2c4bb8096d448c6ae64f8ff90ab0ff463
