+++ src/taipy/core/cycle/_cycle_model.py
@@ -27,7 +27,6 @@ class _CycleModel:
     creation_date: str
     start_date: str
     end_date: str
-    version: str
 
     def to_dict(self) -> Dict[str, Any]:
         return {**dataclasses.asdict(self), "frequency": repr(self.frequency)}
@@ -42,5 +41,4 @@ class _CycleModel:
             creation_date=data["creation_date"],
             start_date=data["start_date"],
             end_date=data["end_date"],
-            version=data["version"],
         )
